Add files via upload
This small work is an exercise of Google Advanced Data Analytics training which I did to understand the Naive Bayes. This exercise let me understand about the importance of assessing the model not only from solely the accuracy score but also from precision, recall and F1 score. The confusion matrix is useful to give the overall detail about the output of the model.